Implement SolveSolos().
This solver method will solve all cells in the puzzle which can be
determined by simple elimination.  We loop over the puzzle and for each
unknown cell, find the set of possible candidate glyphs.  If there is
only one candidate glyph, we set it as known and keep going.  As long as
we have discovered at least one cell in this way, we iterate over the
entire puzzle again.  The method only stops working when it detects no
cells solved in a complete pass, or all cells have been solved.

This method will generally be able to fully solve puzzles which are
labelled 'Easy', i.e., do not require use of advanced solving techniques
such as hidden groups, X-wing, swordfish, jellyfish, XYZ, etc.