Make SolveSolos concurrent.
On each pass, SolveSolos now launches a fleet of goroutines, one for
each unknown cell in the puzzle.  Each goroutine will attempt to solve
that one cell, and returns whether it succeeded on a boolean channel.

As before, SolveSolos stops when it has either solved the puzzle, or
detects that it has not solved any cells on a full pass.