Add basic solution generation functions.
The public function GenerateSolution() produces a valid, complete sudoku
solution (a puzzle with no unknown cells) by repeatedly calling
attemptSolution() until it gets a valid solution.

attemptSolution() works by randomly seeding three rows with values and
then solving the puzzle using the same combination of logical
elimination and guesswork as found in the solver routines.

attemptSolution is written to be used concurrently, but in this commit
GenerateSolution just executes it linearly.  This is low-hanging fruit
for future improvement.  GenerateSolution could be extended to maintain
a pool of attemptSolution workers.