Refactor attemptSolution out into two methods.
The code of attemptSolution was pretty lengthy and not easy to unit test
effectively.  I decided to crack it open and put the seeding logic into
its own function and make it a method of Puzzle.

Likewise, AttemptSolution is now a method of Puzzle and is exported.
This takes an already seeded puzzle and tries to find a solution for it
with some randomness thrown into the mix.