Attempt to resolve an intermittent failure in SolveEasy().
This method was occasionally returning -1 for the number of unknown
cells remaining in the puzzle.  My theory is that this is due to two
goroutines both reporting success on solving the same cell.  The
solution was to do a final count of the number of unknowns remaining at
the bottom of the method before returning.