Perform Clear() at the start of the loop in GenerateSolution().
Technically this is not necessary, Clear() can be performed at the
bottom of the loop (as previously) but this results in a report of
non-coverage of the Clear() call except in the rare case that the first
attempt is a bust.

This tiny performance penalty is an easy price to pay for repeatably
total test coverage.