Add Mask type and method ApplyMask() for puzzle.
The intention is that the Mask object (a two-dimensional boolean array)
will have true values to indicate a "clue" (revealed) cell in the
puzzle, and false values to indicate a "hidden" (concealed) cell.

In this way it will be possible to create puzzles of varying difficulty
from a single solution grid, by applying different masks to it.