Change internal structure of Puzzle to a one-dimensional array.
Rather than a two-dimensional 9x9 array, I have converted the structure
of Puzzle to a one-dimensional array of 81 cells.  This is less
intuitive from a human reading point of view but (hopefully) more
efficient for the program.  Many functions were simplified after this
change, so that's a positive indication at least.